# server/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "企业 IoT 设备管理平台接口文档（设备 / 产品 / 物模型 / 规则 / 告警 / OTA / 开放平台）",
        "title": "KK 物联云 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页查询当前账号可见的产品，附带每产品设备数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "产品"
                ],
                "summary": "产品列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "名称模糊关键字",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Resp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Resp"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "新建设备产品（自动生成 ProductKey；一型一密时生成 ProductSecret）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "产品"
                ],
                "summary": "创建产品",
                "parameters": [
                    {
                        "description": "产品信息 {name, protocol, accessMode, secretMode, pollInterval, description, frameMode, ...}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Resp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Resp"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定产品（含设备数），需有查看权限",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "产品"
                ],
                "summary": "产品详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "产品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Resp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Resp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}